fix: moved logic to hooks
